server/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6

DEFAULT_IMAGE=https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcR8UEWZRFZx3O-Ctpn0NErgGlOrXSkABwunsg&usqp=CAU
SECRET_KEY=PROWLERCARHAHA
DATABASE_URI=postgresql://edstutorialbase_user:eOJuCYrb30ifeG5F5exCFX7gmBAp9X0L@dpg-cirbe45ph6ev5r881t5g-a.oregon-postgres.render.com/edstutorialbase